+++ packages/api/src/controller/books.js
@@ -1,9 +1,13 @@
 import BookService from '../service/books';
+import PermissionService from "../service/permissions";
+import PERMISSION from "../repository/permissions";
 
 const getBooks = {
   method: 'GET',
   path: '/books',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.BOOK_READ);
+
     return await BookService.getBooks();
   },
   options: {
@@ -15,9 +19,11 @@ const editBook = {
   method: 'PUT',
   path: '/books/{id}',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.BOOK_EDIT);
+
     const id = parseInt(request.params.id, 10);
 
-    return await BookService.editBook(id, request.payload);
+    return await BookService.editBook(id, request.payload.input, request.payload.sender_id);
   },
 };
 
@@ -25,6 +31,8 @@ const deleteBook = {
   method: 'DELETE',
   path: '/books/{id}',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.BOOK_DELETE);
+
     return await BookService.deleteBook(request.params.id);
   },
 };
@@ -33,6 +41,8 @@ const createBook = {
   method: 'POST',
   path: '/books',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.BOOK_CREATE);
+
     return await BookService.createBook(request.payload.input, request.payload.sender_id);
   },
 };
@@ -41,6 +51,8 @@ const getBookCategories = {
   method: 'GET',
   path: '/bookCategories',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.BOOK_READ);
+
     return await BookService.getBookCategories();
   },
   options: {
@@ -52,6 +64,8 @@ const editBookCategory = {
   method: 'PUT',
   path: '/bookCategories/{id}',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.BOOK_EDIT);
+
     const id = parseInt(request.params.id, 10);
 
     return await BookService.editBookCategory(id, request.payload);
@@ -62,6 +76,8 @@ const deleteBookCategory = {
   method: 'DELETE',
   path: '/bookCategories/{id}',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.BOOK_DELETE);
+
     return await BookService.deleteBookCategory(request.params.id);
   },
 };
@@ -70,7 +86,9 @@ const createBookCategory = {
   method: 'POST',
   path: '/bookCategories',
   handler: async function (request, h) {
-    return await BookService.createBookCategory(request.payload.input, request.payload.sender_id);
+    await PermissionService.requiresPermission(request, PERMISSION.BOOK_CREATE);
+
+    return await BookService.createBookCategory(request.payload);
   },
 };
