+++ packages/api/src/controller/communityConfig.js
@@ -1,9 +1,13 @@
 import CommunityConfigService from '../service/communityConfig';
+import PermissionService from "../service/permissions";
+import PERMISSION from "../repository/permissions";
 
 const getConfig = {
   method: 'GET',
   path: '/communityConfig',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.CONFIG_READ);
+
     return await CommunityConfigService.getConfig();
   },
   options: {
@@ -15,6 +19,8 @@ const editConfig = {
   method: 'PUT',
   path: '/communityConfig/{key}',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.CONFIG_EDIT);
+
     return await CommunityConfigService.editConfig(request.params.key, request.payload);
   },
 };
@@ -23,6 +29,8 @@ const deleteConfig = {
   method: 'DELETE',
   path: '/communityConfig/{key}',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.CONFIG_EDIT);
+
     return await CommunityConfigService.deleteConfig(request.params.key);
   },
 };
@@ -31,6 +39,8 @@ const createConfig = {
   method: 'POST',
   path: '/communityConfig/{key}',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.CONFIG_EDIT);
+
     return await CommunityConfigService.createConfig(request.params.key, request.payload);
   },
 };
