+++ packages/api/src/controller/theme.js
@@ -1,9 +1,13 @@
 import ThemeService from '../service/theme';
+import PermissionService from "../service/permissions";
+import PERMISSION from "../repository/permissions";
 
 const getTheme = {
   method: 'GET',
   path: '/theme',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.THEME_VIEW);
+
     return await ThemeService.getTheme();
   },
 };
@@ -12,9 +16,11 @@ const editTheme = {
   method: 'PUT',
   path: '/theme/{id}',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.THEME_EDIT);
+
     const id = parseInt(request.params.id, 10);
 
-    return await ThemeService.editTheme(id, request.payload);
+    return await ThemeService.editTheme(id, request.payload.input, request.payload.sender_id);
   },
 };
 
@@ -22,6 +28,8 @@ const deleteTheme = {
   method: 'DELETE',
   path: '/theme/{id}',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.THEME_EDIT);
+
     return await ThemeService.deleteTheme(request.params.id);
   },
 };
@@ -30,6 +38,8 @@ const createTheme = {
   method: 'POST',
   path: '/theme',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.THEME_EDIT);
+
     return await ThemeService.createTheme(request.payload.input, request.payload.sender_id);
   },
 };
@@ -38,6 +48,8 @@ const getUserThemes = {
   method: 'GET',
   path: '/users/:userId/theme',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.USER_THEME_READ);
+
     return await ThemeService.getUserTheme(request.params.id);
   },
 };
@@ -46,6 +58,8 @@ const setUserTheme = {
   method: 'PUT',
   path: '/users/:userId/theme/:themeId',
   handler: async function (request, h) {
+    await PermissionService.requiresPermission(request, PERMISSION.USER_THEME_EDIT);
+
     return await ThemeService.setUserTheme(
       request.params.userId,
       request.params.themeId,
